pyproject.toml
[tool.poetry]
name = "knowledgeminer"
version = "0.1.0"
description = ""
authors = ["arsalangundroo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.9.34"
langchain = "^0.1.2"
llmsherpa = "^0.1.4"
chromadb = "^0.4.22"
faiss-cpu = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
